+++ packages/fast-components-react-base/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.1](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components-react-base@4.5.0...@microsoft/fast-components-react-base@4.5.1) (2019-05-17)
+
+**Note:** Version bump only for package @microsoft/fast-components-react-base
+
+
+
+
+
+# [4.5.0](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components-react-base@4.4.2...@microsoft/fast-components-react-base@4.5.0) (2019-05-17)
+
+
+### Bug Fixes
+
+* update slider to support density and other minor style polish ([#1739](https://github.com/Microsoft/fast-dna/issues/1739)) ([f3a1ba0](https://github.com/Microsoft/fast-dna/commit/f3a1ba0))
+
+
+### Features
+
+* adding value attribute to radio and checkbox ([#1745](https://github.com/Microsoft/fast-dna/issues/1745)) ([4b476f6](https://github.com/Microsoft/fast-dna/commit/4b476f6))
+
+
+
+
+
 ## [4.4.2](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components-react-base@4.4.1...@microsoft/fast-components-react-base@4.4.2) (2019-05-08)